Tidy up VM logic
