+++ WalletWasabi.Fluent/ViewModels/CoinSelection/Columns/IsSelectedThreeStateViewModel.cs
@@ -1,6 +1,7 @@
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
 using ReactiveUI;
+using WalletWasabi.Fluent.ViewModels.CoinSelection.Model;
 
 namespace WalletWasabi.Fluent.ViewModels.CoinSelection.Columns;
 
@@ -26,38 +27,5 @@ public partial class IsSelectedThreeStateViewModel : ViewModelBase, IDisposable
 	public void Dispose()
 	{
 		_disposable.Dispose();
-	}
-}
-
-public static class ConvertSelection
-{
-	public static bool? From(SelectionState state)
-	{
-		if (state == SelectionState.False)
-		{
-			return false;
-		}
-
-		if (state == SelectionState.True)
-		{
-			return true;
-		}
-
-		return null;
-	}
-
-	public static SelectionState To(bool? state)
-	{
-		if (state == true)
-		{
-			return SelectionState.True;
-		}
-
-		if (state == false)
-		{
-			return SelectionState.False;
-		}
-
-		return SelectionState.Partial;
 	}
 }
