+++ WalletWasabi.Fluent/ViewModels/CoinSelection/LabelBasedCoinSelectionViewModel.cs
@@ -8,6 +8,7 @@ using DynamicData;
 using ReactiveUI;
 using WalletWasabi.Fluent.Extensions;
 using WalletWasabi.Fluent.ViewModels.CoinSelection.Columns;
+using WalletWasabi.Fluent.ViewModels.CoinSelection.Model;
 using WalletWasabi.Fluent.ViewModels.Wallets.Advanced.WalletCoins;
 
 namespace WalletWasabi.Fluent.ViewModels.CoinSelection;
@@ -32,7 +33,7 @@ public partial class LabelBasedCoinSelectionViewModel : ViewModelBase, IDisposab
 			.TransformWithInlineUpdate(
 				group =>
 				{
-					var coinGroup = new CoinGroupViewModel(group.Key, group.Cache);
+					var coinGroup = new CoinGroupViewModel(group.Key, group.Cache.Connect());
 					return new TreeNode(coinGroup, coinGroup.Items.Select(x => new TreeNode(x)));
 				})
 			.Filter(filterPredicate)
@@ -62,7 +63,7 @@ public partial class LabelBasedCoinSelectionViewModel : ViewModelBase, IDisposab
 				return true;
 			}
 
-			if (tn.Value is CoinGroupViewModel cg)
+			if (tn.Value is Model.CoinGroupViewModel cg)
 			{
 				var containsLabel = cg.Labels.Any(s => s.Contains(text, StringComparison.InvariantCultureIgnoreCase));
 				return containsLabel;
