+++ WalletWasabi.Fluent/ViewModels/CoinSelection/Model/ISelectable.cs
@@ -1,18 +1,6 @@
-namespace WalletWasabi.Fluent.ViewModels.CoinSelection;
+namespace WalletWasabi.Fluent.ViewModels.CoinSelection.Model;
 
 public interface ISelectable
 {
 	public bool IsSelected { get; set; }
-}
-
-public interface IThreeState
-{
-	public SelectionState SelectionState { get; set; }
-}
-
-public enum SelectionState
-{
-	False,
-	True,
-	Partial,
 }
