+++ WalletWasabi.Fluent/ViewModels/CoinSelection/Model/IThreeState.cs
@@ -1,0 +1,6 @@
+namespace WalletWasabi.Fluent.ViewModels.CoinSelection.Model;
+
+public interface IThreeState
+{
+	public SelectionState SelectionState { get; set; }
+}
