+++ WalletWasabi.Fluent/ViewModels/CoinSelection/Model/SelectionState.cs
@@ -1,0 +1,8 @@
+namespace WalletWasabi.Fluent.ViewModels.CoinSelection.Model;
+
+public enum SelectionState
+{
+	False,
+	True,
+	Partial,
+}
